[actions] Fix incorrect handling of CPU tuning differences
This fixes two bugs - one major and one minor.
(1) Major bug: when we had an IR with the same set of CPU tuning flags as the baseline project's IR, then we would fail to initialize the CPU tuning flags, leading to a false build.
(2) Minor bug: when the file had a different hash than the baseline project, then we should apply the CPU tuning flags immediately.

Currently, we don't detect identical groups